decompose task to smaller files
